refactor(recall): consolidate recall quality policy and share scope helpers
Folds the pure recall quality filter into lifecycle/recall.ts so the
recall core owns query composition, fetch, quality filtering, and
rendering. Tag-scope policy now lives once in operations/memory-scope.ts
(scopeTagsForBank + composeScopedTagFilter), used by the explicit
recall/reflect operations and covered by deterministic tests. Session
memory gating already flows through the shared session-memory-meta
module, so no extra helper was added.

Refs #414
