+++ extensions/operations/memory-recall-operations.ts
@@ -1,5 +1,5 @@
 import type { MemoryOperationsDeps } from "./memory-operation-types.js";
-import { recallScopeTags } from "../banks/banking.js";
+import { composeScopedTagFilter, scopeTagsForBank } from "./memory-scope.js";
 import { resolveOperationBank } from "../banks/bank-selection.js";
 import { readLastRecallSnapshot, resolveLastRecallPath } from "../lifecycle/recall-visibility.js";
 import {
@@ -38,34 +38,6 @@ interface ExplicitReflectFilters extends Omit<ExplicitRecallFilters, "queryTimes
   excludeMentalModelIds?: string[];
 }
 
-function recallTagsForBank(
-  cwd: string,
-  config: ResolvedConfig,
-  projectBankId: string,
-  bankId: string,
-): string[] {
-  return config.banks.user.enabled && bankId === config.banks.user.bankId
-    ? ["source:pi"]
-    : recallScopeTags(cwd);
-}
-
-function scopedTagFilterOptions(scopeTags: string[], filters: ExplicitRecallFilters = {}) {
-  const scopeGroup = { tags: scopeTags, match: "any_strict" } satisfies HindsightTagGroup;
-  const callerGroups = filters.tagGroups ?? [];
-  if (callerGroups.length || filters.tags?.length) {
-    const flatTagGroup = filters.tags?.length
-      ? [
-          {
-            tags: filters.tags,
-            match: filters.tagsMatch ?? "any_strict",
-          } satisfies HindsightTagGroup,
-        ]
-      : [];
-    return { tagGroups: [scopeGroup, ...flatTagGroup, ...callerGroups] };
-  }
-  return { tags: scopeTags, tagsMatch: "any_strict" as const };
-}
-
 export function createRecallOperations(deps: MemoryOperationsDeps) {
   return {
     async recall(
@@ -84,7 +56,7 @@ export function createRecallOperations(deps: MemoryOperationsDeps) {
         config,
         projectBankId: deps.getProjectBankId(),
       });
-      const scopeTags = recallTagsForBank(cwd, config, deps.getProjectBankId(), bankId);
+      const scopeTags = scopeTagsForBank(cwd, config, bankId);
       const result = await deps.getClient().recall(bankId, query, {
         budget: filters.budget ?? config.recall.budget,
         maxTokens: filters.maxTokens ?? config.recall.maxTokens,
@@ -107,7 +79,7 @@ export function createRecallOperations(deps: MemoryOperationsDeps) {
         ...(filters.maxSourceFactsTokens !== undefined
           ? { maxSourceFactsTokens: filters.maxSourceFactsTokens }
           : {}),
-        ...scopedTagFilterOptions(scopeTags, filters),
+        ...composeScopedTagFilter(scopeTags, filters),
         ...(filters.signal ? { signal: filters.signal } : {}),
       });
       return { bankId, result };
@@ -127,7 +99,7 @@ export function createRecallOperations(deps: MemoryOperationsDeps) {
         config,
         projectBankId: deps.getProjectBankId(),
       });
-      const scopeTags = recallTagsForBank(cwd, config, deps.getProjectBankId(), bankId);
+      const scopeTags = scopeTagsForBank(cwd, config, bankId);
       const result = await deps.getClient().reflect(bankId, query, {
         ...(context ? { context } : {}),
         budget: filters.budget ?? config.recall.budget,
@@ -144,7 +116,7 @@ export function createRecallOperations(deps: MemoryOperationsDeps) {
         ...(filters.excludeMentalModelIds
           ? { excludeMentalModelIds: filters.excludeMentalModelIds }
           : {}),
-        ...scopedTagFilterOptions(scopeTags, filters),
+        ...composeScopedTagFilter(scopeTags, filters),
         ...(filters.signal ? { signal: filters.signal } : {}),
       });
       return { bankId, result };
