+++ tests/recall-quality-fixtures.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from "vitest";
 import type { AgentMessage } from "@earendil-works/pi-agent-core";
 import { DEFAULT_CONFIG } from "../extensions/config/config.js";
-import { filterRecallQuality } from "../extensions/lifecycle/recall-quality-policy.js";
+import { filterRecallQuality } from "../extensions/lifecycle/recall.js";
 import { recallForContext, renderRecallBlocks } from "../extensions/lifecycle/recall.js";
 import type { HindsightLikeClient } from "../extensions/types.js";
